+++ src/main/java/com/coursework/controller/UserController.java
@@ -2,63 +2,115 @@ package com.coursework.controller;
 
 import com.coursework.models.User;
 import com.coursework.services.UserService;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 @Controller
 @RequiredArgsConstructor
+@RequestMapping("/admin")
 public class UserController {
     private final UserService userService;
 
-    @GetMapping("/users")
-    public String getAllUsers(Model model) {
-        List<User> users = userService.list();
-        model.addAttribute("users", users);
-        return "userList"; // Предполагается, что у вас есть шаблон с именем "userList"
+    @Controller
+    public class AdminController {
+
+        @GetMapping("/admin/manage-users")
+        public String manageUsers(HttpSession session, Model model) {
+            // Получение информации о пользователе из сессии
+            User loggedInUser = (User) session.getAttribute("loggedInUser");
+            // Проверка, если пользователь не авторизован или не является superAdmin
+            if (loggedInUser == null || !loggedInUser.isSuperAdmin()) {
+                // Перенаправление на страницу с сообщением о запрете доступа
+                return "access-denied";
+            }
+
+            // Получение списка пользователей и передача его в модель для отображения
+            List<User> users = userService.list();
+            model.addAttribute("users", users);
+
+            return "manage-users";
+        }
     }
 
+
+
     @GetMapping("/{id}")
-    public String getUserById(@PathVariable int id, Model model) {
+    public String getUserById(HttpSession session,@PathVariable int id, Model model) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null) {
+
+            return "authorization";
+        }
         User user = userService.getUserById(id);
         model.addAttribute("user", user);
         return "userDetails"; // Предполагается, что у вас есть шаблон с именем "userDetails"
     }
 
     @GetMapping("/create")
-    public String createUserForm(Model model) {
+    public String createUserForm(HttpSession session,Model model) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null) {
+
+            return "authorization";
+        }
         model.addAttribute("user", new User());
         return "createUser"; // Предполагается, что у вас есть шаблон с именем "createUser" для создания нового пользователя
     }
 
     @PostMapping("/create")
-    public String createUser(@ModelAttribute User user) {
+    public String createUser(HttpSession session,@ModelAttribute User user) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null) {
+
+            return "authorization";
+        }
         userService.saveUser(user);
         return "redirect:/users";
     }
 
     @GetMapping("/{id}/edit")
-    public String editUserForm(@PathVariable int id, Model model) {
+    public String editUserForm(HttpSession session,@PathVariable int id, Model model) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null) {
+
+            return "authorization";
+        }
         User user = userService.getUserById(id);
         model.addAttribute("user", user);
         return "editUser"; // Предполагается, что у вас есть шаблон с именем "editUser" для редактирования пользователя
     }
 
     @PostMapping("/{id}/edit")
-    public String editUser(@PathVariable int id, @ModelAttribute User updatedUser) {
+    public String editUser(HttpSession session,@PathVariable int id, @ModelAttribute User updatedUser) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null) {
+
+            return "authorization";
+        }
         updatedUser.setId(id);
         userService.updateUser(updatedUser);
         return "redirect:/users";
     }
 
-    @GetMapping("/{id}/delete")
-    public String deleteUser(@PathVariable int id) {
+    @PostMapping("/{id}/delete")
+    public String deleteUser(HttpSession session,@PathVariable int id) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+
+        if (loggedInUser == null) {
+
+            return "authorization";
+        }
         userService.deleteUser(id);
-        return "redirect:/users";
+        return "redirect:/admin/manage-users";
     }
 }
